scripts/models/requirements.txt
absl-py==0.9.0
astor==0.8.1
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
jieba==0.40
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.16.5
oauthlib==3.1.0
opt-einsum==2.3.2
pandas==0.24.2
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.20.4
scipy==1.2.2
six==1.13.0
sklearn==0.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-serving-api==0.0.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
